+++ index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import ChunkErrorBoundary from './components/ChunkErrorBoundary';
+import { QueryProvider } from './contexts/QueryProvider';
 import { initErrorTracking } from './services/errorTracking';
 
 // Initialize error tracking (Sentry)
@@ -17,7 +18,9 @@ const root = ReactDOM.createRoot(rootElement);
 root.render(
   <React.StrictMode>
     <ChunkErrorBoundary>
-      <App />
+      <QueryProvider>
+        <App />
+      </QueryProvider>
     </ChunkErrorBoundary>
   </React.StrictMode>
 );